Added warning when deleting files
We don’t want people accidentally deleting files, do we?!
Thanks to Habbison for the idea.

+++ resources/assets/js/editor.js
@@ -156,7 +156,9 @@ function addTab(file) {
       deleteButton.innerHTML = '&times;';
       deleteButton.onclick = function(this) {
         let span = this.parentNode;
-        deleteFile(span.dataset.filename);
+        if(confirm('Are you sure you want to delete ' + span.dataset.filename + '?')) {
+          deleteFile(span.dataset.filename);
+        }
       };
       deleteButton.classList.add('smallButton');
       span.appendChild(deleteButton);
